+++ src/auth/application/auth.service.ts
@@ -6,7 +6,7 @@ import {
 } from '@nestjs/common';
 import { JwtService } from '@nestjs/jwt';
 import { ConfigService } from '@nestjs/config';
-import { MailService } from '../../mail/mail.service';
+import { MailService } from '../../mail/application/mail.service';
 import { Cache, CACHE_MANAGER } from '@nestjs/cache-manager';
 import { InjectRepository } from '@nestjs/typeorm';
 import { RoleEntity } from '../../user/infrastructure/entity/roles.entity';
